requirements.txt
pillow
whitenoise
psycopg2-binary
bootstrap4==0.1.0
decorator==4.4.2
Django==3.1.1
django-ads==1.1.1
django-appconf==1.0.4
django-bootstrap3==14.2.0
django-bootstrap4==2.3.1
django-braces==1.14.0
django-classy-tags==2.0.0
django-embed-video==1.4.0
django-js-reverse==0.9.1
django-sekizai==2.0.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
gunicorn==20.0.4
humanize==3.2.0
json5==0.9.5
jsonschema==3.2.0
lazy-object-proxy==1.4.3
libarchive-c==2.9
Markdown==3.3.3
MarkupSafe==1.1.1
mysqlclient==2.0.2
protobuf==3.14.0
python-dateutil==2.8.1
requests-oauthlib==1.3.0
urllib3==1.25.9
dj-database-url
